Added Agent class + revised dependencies
Now using Agent class which holds id, velocity and direction, instead of using HashMaps for keeping track of velocity and direction.
Made changes to all parts of the code that interacted with the former HashMaps. There may be changes still needed to be made.
HashMap structure of agentLocations in MAPFState (HashMap<Agent, Integer>) may be changed if needed.