Re-did setfgh() and assignfgh() methods to only set and assign h
Each vertex can only have a heuristic (h) calculated before-hand. g is instead tied to an agent and depends on how many time steps it has taken for the agent to reach the vertex. Thus, g and f are calculated in simulation-time.
fgh-related code is commented out for easy rollback if needed