Modified adjList to distinguish between up and downgoing edges
adjList now has a new data structure, making use of the new class UpDownNeighourList which holds upgoing and downgoing neighbours in separate arraylists. It also has a print function.
This is needed for A* to know what neighbours an agent going in a certain direction is able to take so that it doesn't move backwards.

Also modified the Ramp printAdjList method to correctly print the adjList