Update to upgoing agents
Agents now have a data member boolean passBaysAble, telling if the agent is allowed to enter passing bays. This only has an effect on upgoing agents (Agent constructor only looks at passBaysAble if the agent is upgoing).
Modified A* and ICTS to account for upgoing agents with passBaysAble = true to enter a passing pay.
ICTS now manually removes a neighbouring move to a passing bay for upgoing agents that are prohibited from entering passing bays