Started CBS
Started with solve() method.
Added data members to MAPFScenario and CTNode that keep track of vertex- and edge constraints.
Further developed A* isStateAllowed() method to take any relevant vertex and/or edge constraints into consideration when checking if a proposed move combination is allowed or not. In other words, making A* work with CBS.