Finalised current CBS rollback mechanism
Note. Current mechanism lets CBS keep the ctPrioQueue for the new CBS solve invocation. As per the meeting agenda and implementation backlog notes, this might not be what we want. In that case, do not import ctPrioQueue, it should be rather easy to do.

The same goes for A*, as per the discussions in the notes.

As per this commit, all algorithms should work in regards ot rollback. Prints are a bit awkward at the moment.