Implemented option to prioritise higher priority agents
This applies to all algorithms bar ICTS since the concept of ICTS is based on looking at all agent paths in a pure BFS manner.

Agents must be specified with a higher or ordinary priority.

When invoking a solve() method, the prioritise function is specified as a parameter. True means prioritising higher prio agents, false means looking at all agents as equals.