Created a main scenario with burst of new agents arriving
ICTS outperforms the rest. Seems to be due to how the algorithms perform their search. A* locks in to its previous solutions. ICTS is still quite flexible in terms of how the agents can go (since it only takes path costs into consideration).
Since the algorithms have different search methods, the rollback states may be more or less optimal depending on the scenario, hence the differences in replanning solution qualities